% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape.R
\name{EventHistory}
\alias{EventHistory}
\title{Load cache of past event details}
\usage{
EventHistory(eventName)
}
\arguments{
\item{eventName}{Name of the event, taken from the URL of the event's web page.}
}
\description{
Load cache of past event details
}
