% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{exportData}
\alias{exportData}
\title{Export Data}
\usage{
exportData(data, name)
}
\arguments{
\item{data}{The data you would like to save}

\item{name}{The name of the file}
}
\description{
Exports data to a csv file in the data/exported_data folder of your working
directory.
}
\examples{
exportData(your_data, your_name);
exportData(data, "John SMITH");
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
