% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{fetchAllYourData}
\alias{fetchAllYourData}
\title{Fetch All of Your Data}
\usage{
fetchAllYourData(your_name)
}
\arguments{
\item{your_name}{The name of the person you're interested in viewing. This
must be in the same format as it appears on the website. As #' of March 2025,
that format is 'Firstname LASTNAME'}
}
\value{
all data for the listed runs associated with that individual
}
\description{
Function to pull all data for an individual. Requires a csv file in the 'data'
folder containing a list of all races for a location. The file name should be
the location name, matching the formatting on the Parkrun website (e.g.
'worcester.csv'). Each run number should be on a new line.
}
\examples{
data <- fetchAllYourData('John SMITH');
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
