% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{fetchArchivedData}
\alias{fetchArchivedData}
\title{Fetch Archived Data}
\usage{
fetchArchivedData(your_name)
}
\arguments{
\item{your_name}{The name of the person you're interested in viewing. This
must be in the same format as it appears on the website. As of May 2025,
that format is 'Firstname LASTNAME'}
}
\value{
all archived data for the listed runs associated with that individual
}
\description{
Function to fetch all archived data for an individual from the 'exported_data'
folder in the working directory.
}
\examples{
data <- fetchArchivedData('John SMITH');
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
