% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{fetchRaceDataForIndividual}
\alias{fetchRaceDataForIndividual}
\title{Fetch Race Data for Individual}
\usage{
fetchRaceDataForIndividual(location, race_number, your_name)
}
\arguments{
\item{location}{The Parkrun location. Usually the location (e.g. 'ludlow')}

\item{race_number}{The race number. An integer (e.g. 243)}

\item{your_name}{The name of the person you're interested in viewing. This
must be in the same format as it appears on the website. As #' of March 2025,
that format is 'Firstname LASTNAME'}
}
\value{
the data for that individual
}
\description{
Helper function to scrape data from the Parkrun website for a particular race
and individual. Saves the scraped line to a file.
}
\examples{
data <- fetchRaceDataForIndividual('worcester', 431, 'John SMITH');
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
