% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape.R
\name{GetResults}
\alias{GetResults}
\title{Obtain parkrun results from cache, scraping if not available}
\usage{
GetResults(eventName, runSeqNumber)
}
\arguments{
\item{runSeqNumber}{Integer specifying the event number (see details).}
}
\description{
Obtain parkrun results from cache, scraping if not available
}
\author{
Martin R. Smith
}
