% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrun-package.R
\docType{package}
\name{parkrun-package}
\alias{parkrun}
\alias{parkrun-package}
\title{parkrun: Scrape and Analyse Parkrun Data}
\description{
This package contains scripts that allow Parkrun data to be scraped from the website, converted into a table, and analysed to provide further insights.
}
\keyword{internal}
