% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{plotPositionByTime}
\alias{plotPositionByTime}
\title{Plot Position by Time}
\usage{
plotPositionByTime(data)
}
\arguments{
\item{data}{The data to be plotted. This must be the processed form, not the
raw scraped form}
}
\value{
A plot of the data
}
\description{
Plots the finishing position against completion time
}
\examples{
plot <- plotPositionByTime(data);
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
