% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{processData}
\alias{processData}
\title{Process Data}
\usage{
processData(data, your_name = NULL)
}
\arguments{
\item{data}{The data that needs processing}

\item{your_name}{The name of the person you're interested in viewing, if
applicable. This must be in the same format as it appears on the website. As
of May 2025, that format is 'Firstname LASTNAME'}
}
\value{
a data frame containing the processed data ready for analysis
}
\description{
Processes the data to transform the scraped information into a useable format.
Regular expressions are used to pull out the key data, and any necessary type
conversions are performed. Speed and pace are calculated.
}
\examples{
data <- processData(data, 'John SMITH');
data <- processData(data)
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
