% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape.R
\name{ScrapeEventHistory}
\alias{ScrapeEventHistory}
\title{Start an index file by scraping the dates of previous events}
\usage{
ScrapeEventHistory(eventName,
  eventIndex = paste0(EventDirectory(eventName), "/index.txt"),
  write = TRUE)
}
\arguments{
\item{eventIndex}{Path to \code{index.txt} file containing details of each event}
}
\description{
Start an index file by scraping the dates of previous events
}
