% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunscraper.R
\name{scrapeRaceDate}
\alias{scrapeRaceDate}
\title{Scrape Race Date}
\usage{
scrapeRaceDate(event, race_number)
}
\arguments{
\item{event}{The event name (usually the Parkrun location. e.g. 'worcester')}

\item{race_number}{The race number (an integer. e.g. 431)}
}
\value{
the date of the race
}
\description{
Scrapes the date that a particular race took place
}
\examples{
date <- scrapeRaceDate('worcester', 620);
date <- scrapeRaceDate('ludlow', 65);
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
