% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{SecondsToMinutes}
\alias{SecondsToMinutes}
\title{Convert seconds to minutes}
\usage{
SecondsToMinutes(s)
}
\arguments{
\item{s}{Number of seconds}
}
\value{
A character vector giving a time in minutes, e.g. "61:04".
}
\description{
Convert seconds to minutes
}
\author{
Martin R. Smith
}
