% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scrape.R
\name{SummariseResults}
\alias{SummariseResults}
\title{Summarise parkrun results}
\usage{
SummariseResults(results, athlete = NA, labelSpeed = FALSE)
}
\arguments{
\item{results}{Data frame of results, perhaps generated by \link{ScrapeResults}}

\item{athlete}{Integer vector specifying athlete(s) whose speeds should be indicated on the plot}
}
\description{
Summarise parkrun results
}
\author{
Martin R. Smith
}
