% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkrunplots.R
\name{plotPositionByTime}
\alias{plotPositionByTime}
\title{Plot Position by Time}
\usage{
plotPositionByTime(data, filter_by = NULL)
}
\arguments{
\item{data}{The data to be plotted. This must be the processed form, not the
raw scraped form}

\item{filter_by}{A way to filter the dataset. Options are currently 'sex' and
'category'}
}
\value{
A plot of the data
}
\description{
Plots the finishing position against completion time
}
\examples{
plot <- plotPositionByTime(data);
plot <- plotPositionByTime(data, 'sex')
plotPositionByTime(data, 'category')
}
\author{
Sophie Edgar-Andrews (github @sophieed)
}
